Additional changes requested from CR. Float field for deliverable, current_balance, shipment_volume. Metal Type Order sorting. Provide "Others" selection in the Customer dropdown.

+++ group-metal-management/gmm-common/src/main/java/com/matthey/pmm/gmm/Balance.java
@@ -13,9 +13,9 @@ public interface Balance {
     
     String customer();
     
-    Integer currentBalance();
+    Double currentBalance();
     
-    Integer shipmentVolume();
+    Double shipmentVolume();
     
     @Nullable
     Integer shipmentWindow();
