+++ group-metal-management/gmm-endur-connector/src/main/java/com/matthey/pmm/gmm/ForecastKey.java
@@ -25,7 +25,7 @@ public interface ForecastKey {
     @Nullable
     String comments();
     
-    Integer deliverable();
+    Double deliverable();
     
     @Nullable
     String createTime();
